+++ src/component/component.cpp
@@ -1,8 +1,0 @@
-/*
-Motion SGI Emulator
-Copyright (C) 2026 starfrost
-
-Component.cpp: Component method implementation
-*/
-
-#include "component.hpp"