+++ src/component/gpu/juniper/gf2/gf2_fbc.hpp
@@ -5,6 +5,8 @@
     Copyright (c)2026 starfrost
 
     gf2_fbc.hpp: The Frame Buffer Controller, four AMD AM2903 chips running custom SGI Microcode.
+
+    Also incorproates the Bitplane Controller, an interface to VRAM
 */
 
 #pragma once
@@ -15,6 +17,6 @@ namespace Motion
     class GF2FBC : public Component 
     {
     public: 
-        const char* GetName() { return "AMD Am2903 Framebuffer Controller"; }; 
+        const char* GetName() { return "Framebuffer & Bitplane Controller (AMD Am2903)"; }; 
     }; 
 };