+++ docs/changelog.md
@@ -185,6 +185,7 @@ Similar to a .plan file
 
 ## 2026-08-27
 * unaligned ram accesses
+* fixed init order issue causing first instruction to be corrupted
 
 v0.2.0 TODO: